Add tests for iterators
